Add authorization to project link

+++ app/Http/Controllers/Project/ProjectLinkController.php
@@ -4,14 +4,15 @@ namespace App\Http\Controllers\Project;
 
 use App\Project;
 use App\Link;
-use Illuminate\Http\Request;
-use App\Http\Controllers\LinkController;
+use App\Http\Requests\LinkRequest;
+use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Redirect;
 
-class ProjectLinkController extends LinkController
+class ProjectLinkController extends Controller
 {
     public function index(Project $project)
     {
+        $this->authorize('view', $project);
         $links = $project->other_links;
         $editRoute = route('projects.links.edit', ['project' => $project->pid, 'link' => 'LLIINNKK']);
         return view('projects.links.index', compact('project', 'links', 'editRoute'));
@@ -19,14 +20,18 @@ class ProjectLinkController extends LinkController
 
     public function create(Project $project)
     {
+        $this->authorize('create', $project);
         $link = new Link;
         $formAction = route('projects.links.store', ['project' => $project->pid]);
-        return view('links.edit', compact('link', 'formAction'));
+        $parentModel = $project;
+        $deleteRoute = '';
+        return view('links.edit', compact('link', 'formAction', 'parentModel', 'deleteRoute'));
     }
 
-    public function store(Request $request,Project $project)
+    public function store(LinkRequest $request, Project $project)
     {
-        $link = new Link($this->validateValues($request));
+        $this->authorize('view', $project);
+        $link = new Link($request->validated());
         $link->linkable_subtype = Link::SUBTYPE_PROJECT_OTHER;
         $project->links()->save($link);
         return Redirect::route('projects.links.index', ['project' => $project->pid]);
@@ -34,23 +39,23 @@ class ProjectLinkController extends LinkController
 
     public function edit(Project $project, Link $link)
     {
-        $this->validateModelLink($project, $link, Link::SUBTYPE_PROJECT_OTHER);
-        if ($link->linkable_subtype != Link::SUBTYPE_PROJECT_OTHER) abort(404);
+        $this->authorize('view', $project);
         $deleteRoute = route('projects.links.destroy', ['project' => $project->pid, 'link' => $link->pid]);
         $formAction = route('projects.links.update', ['project' => $project->pid, 'link' => $link->pid]);
-        return view('links.edit', compact('link', 'deleteRoute', 'formAction'));
+        $parentModel = $project;
+        return view('links.edit', compact('link', 'deleteRoute', 'formAction', 'parentModel'));
     }
 
-    public function update(Request $request, Project $project, Link $link)
+    public function update(LinkRequest $request, Project $project, Link $link)
     {
-        $this->validateModelLink($project, $link, Link::SUBTYPE_PROJECT_OTHER);
-        $link->update($this->validateValues($request));
+        $this->authorize('update', $project);
+        $link->update($request->validated());
         return Redirect::route('projects.links.index', ['project' => $project->pid]);
     }
 
     public function destroy(Project $project, Link $link)
     {
-        $this->validateModelLink($project, $link, Link::SUBTYPE_PROJECT_OTHER);
+        $this->authorize('delete', $project);
         $link->delete();
         return Redirect::route('projects.links.index', ['project' => $project->pid]);
     }
